separating into components
